# docs/swagger.yaml
components:
  schemas:
...
    dto.BoardDetailResponse:
      properties:
        columns:
          items:
            $ref: '#/components/schemas/dto.ColumnDetailResponse'
          type: array
          uniqueItems: false
        created_at:
          type: string
        id:
          type: string
        name:
          type: string
        position:
          type: integer
        updated_at:
          type: string
        workspace_id:
          type: string
      type: object
    dto.BoardResponse:
      properties:
        created_at:
          type: string
        id:
          type: string
        name:
          type: string
        position:
          type: integer
        updated_at:
          type: string
        workspace_id:
          type: string
      type: object
...
    dto.ColumnDetailResponse:
      properties:
        board_id:
          type: string
        created_at:
          type: string
        id:
          type: string
        name:
          type: string
        position:
          type: integer
        task_count:
          type: integer
        tasks:
          items:
            $ref: '#/components/schemas/dto.TaskResponse'
          type: array
          uniqueItems: false
        updated_at:
          type: string
      type: object
    dto.CreateBoardRequest:
      properties:
        name:
          type: string
      type: object
    dto.CreateColumnRequest:
      properties:
        name:
          type: string
      type: object
...
    dto.TaskAssigneeResponse:
      properties:
        avatar_key:
          type: string
        display_name:
          type: string
        id:
          type: string
        login:
          type: string
      type: object
    dto.TaskResponse:
      properties:
        assignee:
          $ref: '#/components/schemas/dto.TaskAssigneeResponse'
        column_id:
          type: string
        created_at:
          type: string
        description:
          type: string
        due_date:
          type: string
        id:
          type: string
        position:
          type: integer
        priority:
          type: string
        title:
          type: string
        updated_at:
          type: string
      type: object
...
    dto.UpdateBoardRequest:
      properties:
        name:
          type: string
      type: object
    dto.UpdateColumnRequest:
      properties:
        name:
          type: string
        position:
          type: integer
      type: object
...
info:
  description: Collaborative task management API for development teams.
  title: Tandem API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /api/v1/workspaces/{id}/boards:
    get:
      parameters:
      - description: Workspace ID
        in: path
        name: wsId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/dto.BoardResponse'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden
      security:
      - BearerAuth: []
      summary: List workspace boards
      tags:
      - boards
    post:
      parameters:
      - description: Workspace ID
        in: path
        name: wsId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.CreateBoardRequest'
                description: Board payload
                summary: request
        description: Board payload
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.BoardResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Create a board
      tags:
      - boards
  /api/v1/workspaces/{id}/boards/{boardId}:
    delete:
      parameters:
      - description: Workspace ID
        in: path
        name: wsId
        required: true
        schema:
          type: string
      - description: Board ID
        in: path
        name: boardId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      security:
      - BearerAuth: []
      summary: Delete a board
      tags:
      - boards
    get:
      parameters:
      - description: Workspace ID
        in: path
        name: wsId
        required: true
        schema:
          type: string
      - description: Board ID
        in: path
        name: boardId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.BoardDetailResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get a board
      tags:
      - boards
    patch:
      parameters:
      - description: Workspace ID
        in: path
        name: wsId
        required: true
        schema:
          type: string
      - description: Board ID
        in: path
        name: boardId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.UpdateBoardRequest'
                description: Board payload
                summary: request
        description: Board payload
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.BoardResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      security:
      - BearerAuth: []
      summary: Update a board
      tags:
      - boards
  /api/v1/workspaces/{id}/boards/{boardId}/columns:
    post:
      parameters:
      - description: Workspace ID
        in: path
        name: wsId
        required: true
        schema:
          type: string
      - description: Board ID
        in: path
        name: boardId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.CreateColumnRequest'
                description: Column payload
                summary: request
        description: Column payload
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ColumnDetailResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      security:
      - BearerAuth: []
      summary: Create a column
      tags:
      - columns
  /api/v1/workspaces/{id}/boards/{boardId}/columns/{columnId}:
    delete:
      parameters:
      - description: Workspace ID
        in: path
        name: wsId
        required: true
        schema:
          type: string
      - description: Board ID
        in: path
        name: boardId
        required: true
        schema:
          type: string
      - description: Column ID
        in: path
        name: columnId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      security:
      - BearerAuth: []
      summary: Delete a column
      tags:
      - columns
    patch:
      parameters:
      - description: Workspace ID
        in: path
        name: wsId
        required: true
        schema:
          type: string
      - description: Board ID
        in: path
        name: boardId
        required: true
        schema:
          type: string
      - description: Column ID
        in: path
        name: columnId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/dto.UpdateColumnRequest'
                description: Column payload
                summary: request
        description: Column payload
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ColumnDetailResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      security:
      - BearerAuth: []
      summary: Update a column
      tags:
      - columns
...
servers:
- url: http://localhost:8080/